adding some redundancy in saving scores on the go

+++ utils.js
@@ -2,7 +2,15 @@ const fs = require('fs');
 
 const storeData = (data, path) => {
     try {
-        fs.writeFileSync(path, JSON.stringify(data, null, 4));
+        if( overwrite ){
+          fs.writeFileSync(path, JSON.stringify(data, null, 4));
+        } else {
+          let loaded =loadData(path);
+          if( loaded != undefined ) loaded = JSON.parse(loaded);
+          data.forEach(e => loaded.push(data));
+          fs.writeFileSync(path, JSON.stringify(loaded, null, 4));
+        }
+        
     } catch (err) {
         console.error(err)
     }
